+++ apps/opik-frontend/src/components/pages/ConfigurationPage/AlertsTab/AddEditAlertPage/EventTriggers.tsx
@@ -1,8 +1,7 @@
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import { Path, useFieldArray, UseFormReturn } from "react-hook-form";
 import { CircleHelp, ExternalLink, Plus, WebhookIcon, X } from "lucide-react";
 import get from "lodash/get";
-import { keepPreviousData } from "@tanstack/react-query";
 
 import { Label } from "@/components/ui/label";
 import {
@@ -25,10 +24,7 @@ import { Separator } from "@/components/ui/separator";
 import { Input } from "@/components/ui/input";
 import SelectBox from "@/components/shared/SelectBox/SelectBox";
 import ProjectsSelectBox from "@/components/pages-shared/automations/ProjectsSelectBox";
-import LoadableSelectBox from "@/components/shared/LoadableSelectBox/LoadableSelectBox";
-import useFeedbackDefinitionsList from "@/api/feedback-definitions/useFeedbackDefinitionsList";
-import useTracesFeedbackScoresNames from "@/api/traces/useTracesFeedbackScoresNames";
-import useThreadsFeedbackScoresNames from "@/api/traces/useThreadsFeedbackScoresNames";
+import AlertFeedbackScoresSelect from "@/components/shared/AlertFeedbackScoresSelect/AlertFeedbackScoresSelect";
 import { DropdownOption } from "@/types/shared";
 import { AlertFormType } from "./schema";
 import { TRIGGER_CONFIG } from "./helpers";
@@ -36,7 +32,6 @@ import { ALERT_EVENT_TYPE } from "@/types/alerts";
 import { useIsFeatureEnabled } from "@/components/feature-toggles-provider";
 import { FeatureToggleKeys } from "@/types/feature-toggles";
 import { cn } from "@/lib/utils";
-import useAppStore from "@/store/AppStore";
 
 type EventTriggersProps = {
   form: UseFormReturn<AlertFormType>;
@@ -60,126 +55,6 @@ const OPERATOR_OPTIONS: DropdownOption<string>[] = [
   { label: ">", value: ">" },
   { label: "<", value: "<" },
 ];
-
-const DEFAULT_LOADED_FEEDBACK_DEFINITION_ITEMS = 1000;
-
-interface FeedbackScoreNameSelectorProps {
-  value: string;
-  onChange: (value: string) => void;
-  className?: string;
-  eventType: ALERT_EVENT_TYPE;
-}
-
-const FeedbackScoreNameSelector: React.FC<FeedbackScoreNameSelectorProps> = ({
-  value,
-  onChange,
-  className,
-  eventType,
-}) => {
-  const workspaceName = useAppStore((state) => state.activeWorkspaceName);
-  const [isLoadedMore, setIsLoadedMore] = useState(false);
-
-  // Fetch feedback definitions
-  const { data: feedbackDefinitionsData, isLoading: isLoadingDefinitions } =
-    useFeedbackDefinitionsList(
-      {
-        workspaceName,
-        page: 1,
-        size: isLoadedMore ? 10000 : DEFAULT_LOADED_FEEDBACK_DEFINITION_ITEMS,
-      },
-      {
-        placeholderData: keepPreviousData,
-      },
-    );
-
-  // Fetch actual feedback score names based on trigger type
-  const isTraceTrigger =
-    eventType === ALERT_EVENT_TYPE.trace_feedback_score;
-  const isThreadTrigger =
-    eventType === ALERT_EVENT_TYPE.trace_thread_feedback_score;
-
-  const { data: traceScoresData, isLoading: isLoadingTraceScores } =
-    useTracesFeedbackScoresNames(
-      { projectId: undefined },
-      { enabled: isTraceTrigger },
-    );
-
-  const { data: threadScoresData, isLoading: isLoadingThreadScores } =
-    useThreadsFeedbackScoresNames(
-      { projectId: undefined },
-      { enabled: isThreadTrigger },
-    );
-
-  const total = feedbackDefinitionsData?.total ?? 0;
-
-  const loadMoreHandler = useCallback(() => setIsLoadedMore(true), []);
-
-  // Merge feedback definitions and actual feedback score names
-  const options: DropdownOption<string>[] = useMemo(() => {
-    const definitionNames = new Map<
-      string,
-      { name: string; description?: string }
-    >();
-
-    // Add feedback definitions
-    (feedbackDefinitionsData?.content || []).forEach((def) => {
-      definitionNames.set(def.name, {
-        name: def.name,
-        description: def.description,
-      });
-    });
-
-    // Add actual feedback score names from appropriate source
-    const scoreNames = isTraceTrigger
-      ? traceScoresData?.scores || []
-      : isThreadTrigger
-        ? threadScoresData?.scores || []
-        : [];
-
-    scoreNames.forEach((score) => {
-      if (!definitionNames.has(score.name)) {
-        definitionNames.set(score.name, { name: score.name });
-      }
-    });
-
-    // Convert to dropdown options, sorted alphabetically
-    return Array.from(definitionNames.values())
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .map((item) => ({
-        value: item.name,
-        label: item.name,
-        description: item.description,
-      }));
-  }, [
-    feedbackDefinitionsData?.content,
-    traceScoresData?.scores,
-    threadScoresData?.scores,
-    isTraceTrigger,
-    isThreadTrigger,
-  ]);
-
-  const isLoading =
-    isLoadingDefinitions ||
-    (isTraceTrigger && isLoadingTraceScores) ||
-    (isThreadTrigger && isLoadingThreadScores);
-
-  return (
-    <LoadableSelectBox
-      value={value}
-      onChange={onChange}
-      options={options}
-      onLoadMore={
-        total > DEFAULT_LOADED_FEEDBACK_DEFINITION_ITEMS && !isLoadedMore
-          ? loadMoreHandler
-          : undefined
-      }
-      buttonClassName={className}
-      isLoading={isLoading}
-      optionsCount={DEFAULT_LOADED_FEEDBACK_DEFINITION_ITEMS}
-      placeholder="Select score"
-    />
-  );
-};
 
 function getThresholdLabel(eventType: ALERT_EVENT_TYPE): string {
   switch (eventType) {
@@ -337,7 +212,7 @@ const EventTriggers: React.FunctionComponent<EventTriggersProps> = ({
             return (
               <FormItem className="min-w-[150px] flex-1">
                 <FormControl>
-                  <FeedbackScoreNameSelector
+                  <AlertFeedbackScoresSelect
                     value={field.value as string}
                     onChange={field.onChange}
                     eventType={eventType}
